refactor: extract segmentation utilities
